copy LaTeX source from old votable.tex file (does not build)
The body of the votable.tex file that formed the main VOTable document
in volute is copied wholesale to the body of the ivoatex VOTable.tex file,
and most of the headers are copied to the header of VOTable.tex.
The original votable.tex file is removed at this commit.

This version does not build, since the Cgiprint machinery that was
used to produce HTML and PDF output from the same LaTeX source file
is no longer present.