fix VOTable.tex so it builds under ivoatex
Make minimal required changes to VOTable.tex so that it will build
a PDF and HTML file under ivoatex.  The output looks reasonably
healthy and correct at this commit, though I haven't been through
exhaustively to check that there are no unwanted content changes.

The main thing required is removing \ifhtx conditionals that were
used to produce different output in HTML and print context, which
included a fair bit of cosmetic wrangling to make the HTML look nice.
Also remove those commands in the header which do nothing,
either because they are related to HTX processing or because they
were redundant in any case.