adjust margins
Tweak the left margins to make the text more central on the page;
something about the ivoatex setup is probably shifting this sideways
from where it used to be.

The margins are still defined in a somewhat ad-hoc way; it may be
desirable to do something about that (e.g. remove non-standard
settings altogether) for consistency with other documents, but leave
it for now since some of the pages would need adjustment to cope
with a smaller width.