replace citations using BibTeX
Replace the hard-coded "References" section with a \bibliography
command and corresponding \cite[tp] commands in the text.
In some cases the wording has been adjusted slightly in accordance
with the ivoatex recommendations; in a couple of other cases the
reference in the text has been removed altogether where it's
not required for understanding the meaning or replaced by a
URL-bearing footnote.

Where possible, references are taken from the standard ivoatex .bib
files, but a new file localrefs.bib has been introduced for
references not covered there.  Possibly entries from here could be
moved into the standard .bib files at some point.