Enabling COOSYS-based epoch propagation.
This entails allowing FIELDref-s and PARAMref-s in COOSYS and defining
role names for the various compontents.

In consequence, COOSYS now is no longer derived from string. Up to now,
writing <COOSYS...>Some random text</COOSYS> was legal.  I am pretty
sure that was not intended.

Also, adding regression tests for the xSYS elements.

If this is agreed upon, we should probably allow a votable:location on
TIMESYS, too.