The default COOSYS/@system is no FK5 rather than the deprecated eq_FK5.
Mark has proposed to go for ICRS
<https://github.com/ivoa-std/VOTable/pull/42#issuecomment-1710034907>, and that
would probably be a reasonable choice, except something like

	<COOSYS equinox="J1992.2"/>

has been legal so far, and with an ICRS default, that would make no sense
(ICRS is only defined for the equator of J2000.0).

Well, such defaults are always evil, and we should probably discourage
relying on them.