Clarify and correct Units section for VOUnits
Although section 4.4 referenced VOUnits as the unit syntax,
the summary and examples in the text did not actually conform
to VOUnits syntax.
I have corrected the text and examples, and clarified the
text to say that unit attributes SHOULD conform to VOUnits.
The use of SHOULD rather than MUST here reflects the consensus
of the discussion on the apps mailing list at
http://mail.ivoa.net/pipermail/apps/2023-May/001576.html

I have *not* updated the examples in the "Possible VOTable extensions"
Appendix A, which contains all manner of abominations alongside
non-VOUnits compliant unit attributes.